v2.1.4

+++ src/main/java/tn/esprit/projetpiback/repository/EvenementRepository.java
@@ -3,5 +3,6 @@ package tn.esprit.projetpiback.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.projetpiback.entites.Evenement;
 
-public interface EvenementRepository extends JpaRepository<Evenement,Long> {
+public interface EvenementRepository extends JpaRepository<Evenement,Integer> {
+
 }
